requirements-mcp.txt
# MCP Server Dependencies
mcp>=1.0.0
asyncio-mqtt>=0.16.0

# Existing dependencies from requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
sqlalchemy==2.0.23
python-multipart==0.0.6
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
boto3==1.35.0

# Additional MCP-specific dependencies
websockets>=12.0
aiofiles>=23.2.0

